backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
asyncpg
redis
pydantic-settings
pyjwt>=2.12.0
cryptography>=46.0.7
requests>=2.33.0
python-multipart>=0.0.26
httpx
pytest>=9.0.3
httpx
ruff
gunicorn
mypy
